Windows build script: remove prior artifacts if present

+++ build-windows.ps1
@@ -9,12 +9,20 @@ fvm flutter build windows
 
 $Root = Get-Location
 cd .\build\windows\runner
-rm -r -Force uspsa-result-viewer
-mkdir uspsa-result-viewer
+
+If(Test-Path uspsa-result-viewer) {
+    rm -r -Force uspsa-result-viewer
+    mkdir uspsa-result-viewer
+}
+
 cp -r -Force .\Release\* uspsa-result-viewer
 cp $Root\data\L2s-Since-2019.json uspsa-result-viewer
 cp $Root\data\Nationals-and-Area-Matches.json uspsa-result-viewer
-rm uspsa-result-viewer.zip
+
+if(Test-Path uspsa-result-viewer.zip) {
+    rm uspsa-result-viewer.zip
+}
+
 Compress-Archive -Path .\uspsa-result-viewer -DestinationPath uspsa-result-viewer.zip -Force
 cd $Root
 cp -Force .\build\windows\runner\uspsa-result-viewer.zip uspsa-result-viewer-$version-windows.zip
